move concurrentExecutor to a utility file
